+++ app/pages/users/all_users_page.php
@@ -1,0 +1,4 @@
+<?php
+
+echo $data['users_table'];
+?>